Fixed warning in v1.0

+++ test/amp/ReduceTest/ReduceTest.cpp
@@ -150,7 +150,7 @@ public:
 template <typename T>
 T generateRandom()
 {
-    float value = rand();
+    double value = rand();
     static bool negate = true;
     if (negate)
     {
